+++ src/main/java/kr/co/co_working/memberTeam/service/MemberTeamService.java
@@ -19,11 +19,8 @@ import java.util.Optional;
 @Service
 @RequiredArgsConstructor
 public class MemberTeamService {
-    private final MemberTeamService service;
-    private final MemberService memberService;
     private final MemberTeamRepository repository;
-    private final MemberRepository memberRepository;
-    private final TeamRepository teamRepository;
+
     /**
      * createMemberTeam : MemberTeam 등록
      * @param member
@@ -101,40 +98,5 @@ public class MemberTeamService {
 
         // 3. MemberTeam 삭제
         repository.delete(memberTeam);
-    }
-
-    @Transactional
-    public void deleteMemberFromTeam(String email, Long teamId) throws NoSuchElementException, Exception {
-        // 1. Team 조회
-        Optional<Team> selectedTeam = teamRepository.findById(teamId);
-
-        // 2. 부재 시 예외 처리
-        if (selectedTeam.isEmpty()) {
-            throw new NoSuchElementException("수정하려는 팀이 존재하지 않습니다. ID : " + teamId);
-        }
-
-        // 3. Member 조회
-        Optional<Member> selectedMember = memberRepository.findById(email);
-
-        // 4. 부재 시 예외 처리
-        if (selectedMember.isEmpty()) {
-            throw new NoSuchElementException("제외하려는 멤버가 존재하지 않습니다. Email : " + email);
-        }
-
-        // 5. 객체 추출
-        Member member = selectedMember.get();
-        Team team = selectedTeam.get();
-
-        // 6. DTO 생성
-        MemberRequestDto.UPDATE memberDto = MemberRequestDto.UPDATE.builder()
-                .email(email)
-                .teamId(null)
-                .build();
-
-        // 7. Member 수정
-        memberService.updateMember(memberDto);
-
-        // 8. MemberTeam 삭제
-        service.deleteMemberTeam(member, team);
     }
 }
